// src/gen/model/v1alpha1KubernetesDiscovery.ts
/**
 * tilt
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.20.2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { V1ObjectMeta } from './v1ObjectMeta';
import { V1alpha1KubernetesDiscoverySpec } from './v1alpha1KubernetesDiscoverySpec';
import { V1alpha1KubernetesDiscoveryStatus } from './v1alpha1KubernetesDiscoveryStatus';

/**
* KubernetesDiscovery
*/
export class V1alpha1KubernetesDiscovery {
    /**
    * APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
    */
    'apiVersion'?: string;
    /**
    * Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
    */
    'kind'?: string;
    'metadata'?: V1ObjectMeta;
    'spec'?: V1alpha1KubernetesDiscoverySpec;
    'status'?: V1alpha1KubernetesDiscoveryStatus;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "apiVersion",
            "baseName": "apiVersion",
            "type": "string"
        },
        {
            "name": "kind",
            "baseName": "kind",
            "type": "string"
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "V1ObjectMeta"
        },
        {
            "name": "spec",
            "baseName": "spec",
            "type": "V1alpha1KubernetesDiscoverySpec"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "V1alpha1KubernetesDiscoveryStatus"
        }    ];

    static getAttributeTypeMap() {
        return V1alpha1KubernetesDiscovery.attributeTypeMap;
    }
}